@@ -1,10 +1,10 @@
 name: signalr_netcore
 description: A flutter client for ASP.NET Core SignalR. The client is able to invoke server side hub functions and to receive method invocations issued by the server. 
-version: 0.1.7+2
+version: 0.1.7+2-nullsafety.0
 homepage: https://github.com/sefidgaran/signalr_client
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
 
@@ -12,7 +12,7 @@ dependencies:
     sdk: flutter
 
   logging: ^1.0.1
-  w3c_event_source: ^1.3.0
+  w3c_event_source: ^1.3.2
 
 dev_dependencies:
   flutter_test:
