@@ -20,7 +20,7 @@ class HandshakeRequestMessage {
 /// private
 class HandshakeResponseMessage {
   // Properties
-  final String error;
+  final String? error;
 
   // Methods
   HandshakeResponseMessage(this.error);
@@ -33,7 +33,7 @@ class HandshakeResponseMessage {
 class ParseHandshakeResponseResult {
   // Properites
   /// Either a string (json) or a Uint8List (binary).
-  final Object remainingData;
+  final Object? remainingData;
 
   /// The HandshakeResponseMessage
   final HandshakeResponseMessage handshakeResponseMessage;
@@ -56,10 +56,10 @@ class HandshakeProtocol {
 
   /// Parse the handshake reponse
   /// data: either a string (json) or a Uint8List (binary) of the handshake response data.
-  ParseHandshakeResponseResult parseHandshakeResponse(Object data) {
+  ParseHandshakeResponseResult parseHandshakeResponse(Object? data) {
     HandshakeResponseMessage responseMessage;
     String messageData;
-    Object remainingData;
+    Object? remainingData;
 
     if (data is Uint8List) {
       // Format is binary but still need to read JSON text from handshake response
@@ -76,7 +76,7 @@ class HandshakeProtocol {
           ? data.sublist(responseLength, data.length)
           : null;
     } else {
-      final String textData = data;
+      final String textData = data as String;
       final separatorIndex =
           textData.indexOf(TextMessageFormat.recordSeparator);
       if (separatorIndex == -1) {
