@@ -11,24 +11,24 @@ import 'utils.dart';
 class ServerSentEventsTransport implements ITransport {
   // Properties
   final SignalRHttpClient _httpClient;
-  final AccessTokenFactory _accessTokenFactory;
+  final AccessTokenFactory? _accessTokenFactory;
 
-  final Logger _logger;
+  final Logger? _logger;
   final bool _logMessageContent;
-  EventSource _eventSource;
-  StreamSubscription<MessageEvent> _eventSourceSub;
-  String _url;
+  EventSource? _eventSource;
+  StreamSubscription<MessageEvent>? _eventSourceSub;
+  String? _url;
 
   @override
-  OnClose onClose;
+  OnClose? onClose;
 
   @override
-  OnReceive onReceive;
+  OnReceive? onReceive;
 
   ServerSentEventsTransport(
       SignalRHttpClient httpClient,
-      AccessTokenFactory accessTokenFactory,
-      Logger logger,
+      AccessTokenFactory? accessTokenFactory,
+      Logger? logger,
       bool logMessageContent)
       : assert(httpClient != null),
         _httpClient = httpClient,
@@ -38,7 +38,7 @@ class ServerSentEventsTransport implements ITransport {
 
   // Methods
   @override
-  Future<void> connect(String url, TransferFormat transferFormat) async {
+  Future<void> connect(String? url, TransferFormat transferFormat) async {
     assert(!isStringEmpty(url));
     assert(transferFormat != null);
     _logger?.finest("(SSE transport) Connecting");
@@ -47,11 +47,12 @@ class ServerSentEventsTransport implements ITransport {
     _url = url;
 
     if (_accessTokenFactory != null) {
-      final token = await _accessTokenFactory();
+      final token = await _accessTokenFactory!();
       if (!isStringEmpty(token)) {
         final encodedToken = Uri.encodeComponent(token);
-        url +=
-            (url.indexOf("?") < 0 ? "?" : "&") + "access_token=$encodedToken";
+        url = url! +
+            (url.indexOf("?") < 0 ? "?" : "&") +
+            "access_token=$encodedToken";
       }
     }
 
@@ -61,22 +62,21 @@ class ServerSentEventsTransport implements ITransport {
           "The Server-Sent Events transport only supports the 'Text' transfer format"));
     }
 
-    _eventSource = EventSource(Uri.parse(url));
+    _eventSource = EventSource(Uri.parse(url!));
 
-    _eventSourceSub = _eventSource.events.listen((MessageEvent event) {
+    _eventSourceSub = _eventSource!.events.listen((MessageEvent event) {
       if (onReceive != null) {
         try {
           //_logger.log(LogLevel.Trace, "(SSE transport) data received. ${getDataDetail(e.data, this.logMessageContent)}.`);
           _logger?.finest("(SSE transport) data received");
-          onReceive(event.data);
+          onReceive!(event.data);
         } catch (error) {
-          _close(error);
           return;
         }
       }
     }, onError: (Object error) {
       if (opened) {
-        _close(error);
+        _close(error as Error?);
       }
     }, onDone: () {
       _close(null);
@@ -84,7 +84,7 @@ class ServerSentEventsTransport implements ITransport {
   }
 
   @override
-  Future<void> send(Object data) async {
+  Future<void> send(Object? data) async {
     if (_eventSource == null) {
       return Future.error(
           new GeneralError("Cannot send until the transport is connected"));
@@ -94,24 +94,24 @@ class ServerSentEventsTransport implements ITransport {
   }
 
   @override
-  Future<void> stop(Error error) {
+  Future<void> stop(Error? error) {
     _close(error);
     return Future.value(null);
   }
 
-  _close(Error error) {
+  _close(Error? error) {
     if (_eventSourceSub != null) {
-      _eventSourceSub.cancel();
+      _eventSourceSub!.cancel();
       _eventSource = null;
 
       if (onClose != null) {
-        Exception ex;
+        Exception? ex;
         if (error != null) {
           ex = (error is Exception)
-              ? error
+              ? error as Exception?
               : new GeneralError(error?.toString());
         }
-        onClose(ex);
+        onClose!(ex);
       }
     }
   }
