@@ -3,21 +3,21 @@ import 'package:logging/logging.dart';
 import 'ihub_protocol.dart';
 import 'signalr_client.dart';
 
-bool isStringEmpty(String value) {
+bool isStringEmpty(String? value) {
   return (value == null) || (value.length == 0);
 }
 
-bool isListEmpty(List value) {
+bool isListEmpty(List? value) {
   return (value == null) || (value.length == 0);
 }
 
 Future<void> sendMessage(
-    Logger logger,
+    Logger? logger,
     String transportName,
     SignalRHttpClient httpClient,
-    String url,
-    AccessTokenFactory accessTokenFactory,
-    Object content,
+    String? url,
+    AccessTokenFactory? accessTokenFactory,
+    Object? content,
     bool logMessageContent) async {
   MessageHeaders headers = MessageHeaders();
   if (accessTokenFactory != null) {
